# openapi/backoffice-api.yaml
openapi: 3.0.0
info:
  title: "PowerFind endpointsx"
  description: >
    PowerFind endpoints
  version: snapshot


paths:
  /backoffice/location:
    get:
      summary: get available locations for powerbanks
      operationId: getLocation
      responses:
        200:
          description: get Locations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
    post:
      summary: add new location for powerbank
      operationId: postLocation
      parameters:
        - name: city
          in: query
          required: true
          schema:
            type: string
        - name: district
          in: query
          required: true
          schema:
            type: string
        - name: address
          in: query
          required: true
          schema:
            type: string
        - name: latitude
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: longitude
          in: query
          required: true
          schema:
            type: number
            format: double
      responses:
        200:
          description: create new location
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditResponse'
        400:
          $ref: "#/components/responses/BadRequest"

  /backoffice/powerbank:
    get:
      summary: get available powerbank for user
      operationId: getPowerbank
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: powerbankId
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: requestedDurationMinutes
          in: query
          required: false
          schema:
            type: integer
          description: Number of minutes the user wants to take the powerbank
      responses:
        200:
          description: get powerbank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditResponse'
        400:
          $ref: "#/components/responses/BadRequest"

    post:
      description: add new powerbank
      operationId: postPowerbank
      parameters:
        - name: model
          in: query
          required: true
          schema:
            type: string
        - name: voltage
          in: query
          required: true
          schema:
            type: string
        - name: health
          in: query
          required: true
          schema:
            type: string
        - name: capacity
          in: query
          required: true
          schema:
            type: string
        - name: location
          in: query
          required: true
          schema:
            type: string
        - name: district
          in: query
          required: true
          schema:
            type: string
      responses:
        200:
          description: add new powerbank
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EditResponse'
        400:
          $ref: "#/components/responses/BadRequest"


components:
  schemas:
    LocationResponse:
      type: array
      items:
        $ref: '#/components/schemas/Location'

    Location:
      type: object
      properties:
        streetAddress:
          type: string
        city:
          type: string
        district:
          type: string

    DeleteResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    EditResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    CreateResponse:
      type: object
      required:
        - message
        - id
      properties:
        message:
          type: string
        id:
          type: string
          format: uuid

    BadRequestResponse:
      required:
        - message
      properties:
        message:
          type: string


  responses:
    BadRequest:
      description: Bad request (400)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestResponse'